+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/fir/firFunction.kt
@@ -2,6 +2,7 @@ package com.javiersc.kotlin.compiler.extensions.fir
 
 import com.javiersc.kotlin.compiler.extensions.common.toCallableId
 import com.javiersc.kotlin.compiler.extensions.common.toName
+import org.jetbrains.kotlin.GeneratedDeclarationKey
 import org.jetbrains.kotlin.descriptors.EffectiveVisibility
 import org.jetbrains.kotlin.descriptors.Modality
 import org.jetbrains.kotlin.descriptors.Visibilities
@@ -10,10 +11,14 @@ import org.jetbrains.kotlin.fir.declarations.FirDeclarationOrigin
 import org.jetbrains.kotlin.fir.declarations.FirFunction
 import org.jetbrains.kotlin.fir.declarations.builder.buildSimpleFunction
 import org.jetbrains.kotlin.fir.declarations.impl.FirResolvedDeclarationStatusImpl
+import org.jetbrains.kotlin.fir.declarations.origin
 import org.jetbrains.kotlin.fir.moduleData
 import org.jetbrains.kotlin.fir.symbols.impl.FirNamedFunctionSymbol
 import org.jetbrains.kotlin.fir.types.impl.FirImplicitNothingTypeRef
 import org.jetbrains.kotlin.name.StandardClassIds
+
+public fun FirSession.nothingFirFunction(key: GeneratedDeclarationKey): FirFunction =
+    nothingFirFunction(key.origin)
 
 public fun FirSession.nothingFirFunction(origin: FirDeclarationOrigin): FirFunction =
     buildSimpleFunction {