+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/fir/firTypeParameterRef.kt
@@ -1,9 +1,11 @@
 package com.javiersc.kotlin.compiler.extensions.fir
 
+import org.jetbrains.kotlin.GeneratedDeclarationKey
 import org.jetbrains.kotlin.fir.FirSession
 import org.jetbrains.kotlin.fir.declarations.FirDeclarationOrigin
 import org.jetbrains.kotlin.fir.declarations.FirTypeParameterRef
 import org.jetbrains.kotlin.fir.declarations.builder.buildTypeParameter
+import org.jetbrains.kotlin.fir.declarations.origin
 import org.jetbrains.kotlin.fir.moduleData
 import org.jetbrains.kotlin.fir.symbols.FirBasedSymbol
 import org.jetbrains.kotlin.fir.symbols.impl.FirTypeParameterSymbol
@@ -11,6 +13,15 @@ import org.jetbrains.kotlin.fir.types.ConeTypeProjection
 import org.jetbrains.kotlin.fir.types.toRegularClassSymbol
 import org.jetbrains.kotlin.fir.types.type
 import org.jetbrains.kotlin.types.Variance
+
+public fun ConeTypeProjection.toFirTypeParameter(
+    session: FirSession,
+    key: GeneratedDeclarationKey,
+    containingDeclarationSymbol: FirBasedSymbol<*>,
+    isReified: Boolean = false,
+    variance: Variance = Variance.INVARIANT,
+): FirTypeParameterRef =
+    toFirTypeParameter(session, key.origin, containingDeclarationSymbol, isReified, variance)
 
 public fun ConeTypeProjection.toFirTypeParameter(
     session: FirSession,