+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/fir/firValueParameter.kt
@@ -1,27 +1,43 @@
 package com.javiersc.kotlin.compiler.extensions.fir
 
+import org.jetbrains.kotlin.GeneratedDeclarationKey
 import org.jetbrains.kotlin.fir.FirSession
 import org.jetbrains.kotlin.fir.declarations.FirContextReceiver
 import org.jetbrains.kotlin.fir.declarations.FirDeclarationOrigin
 import org.jetbrains.kotlin.fir.declarations.FirValueParameter
 import org.jetbrains.kotlin.fir.declarations.builder.FirValueParameterBuilder
 import org.jetbrains.kotlin.fir.declarations.builder.buildValueParameter
+import org.jetbrains.kotlin.fir.declarations.origin
 import org.jetbrains.kotlin.fir.moduleData
-import org.jetbrains.kotlin.fir.symbols.impl.FirClassLikeSymbol
-import org.jetbrains.kotlin.fir.symbols.impl.FirClassSymbol
 import org.jetbrains.kotlin.fir.symbols.impl.FirFunctionSymbol
 import org.jetbrains.kotlin.fir.symbols.impl.FirValueParameterSymbol
+import org.jetbrains.kotlin.fir.types.ConeKotlinType
+import org.jetbrains.kotlin.fir.types.FirResolvedTypeRef
 import org.jetbrains.kotlin.fir.types.FirTypeRef
+import org.jetbrains.kotlin.fir.types.classId
+import org.jetbrains.kotlin.fir.types.coneType
+import org.jetbrains.kotlin.fir.types.toFirResolvedTypeRef
 import org.jetbrains.kotlin.name.Name
 
-public fun FirTypeRef.toValueParameter(
+public fun ConeKotlinType.toValueParameter(
+    session: FirSession,
+    key: GeneratedDeclarationKey,
+    containingFunctionSymbol: FirFunctionSymbol<*>,
+    block: FirValueParameterBuilder.(FirTypeRef) -> Unit = {},
+): FirValueParameter = toValueParameter(session, key.origin, containingFunctionSymbol, block)
+
+public fun ConeKotlinType.toValueParameter(
     session: FirSession,
     origin: FirDeclarationOrigin,
     containingFunctionSymbol: FirFunctionSymbol<*>,
     block: FirValueParameterBuilder.(FirTypeRef) -> Unit = {},
 ): FirValueParameter {
-    val typeRef: FirTypeRef = this
-    val name: Name = typeRef.toClassId(session).shortClassName
+    val coneKotlinType: ConeKotlinType = this
+    val typeRef: FirResolvedTypeRef = coneKotlinType.toFirResolvedTypeRef()
+    val name: Name =
+        checkNotNull(coneKotlinType.classId?.shortClassName) {
+            "`ConeKotlinType` must be a `ConeClassLikeType` in order to find the `classId`"
+        }
     return buildValueParameter {
         this.moduleData = session.moduleData
         this.origin = origin
@@ -36,6 +52,27 @@ public fun FirTypeRef.toValueParameter(
     }
 }
 
+public fun FirTypeRef.toValueParameter(
+    session: FirSession,
+    key: GeneratedDeclarationKey,
+    containingFunctionSymbol: FirFunctionSymbol<*>,
+    block: FirValueParameterBuilder.(FirTypeRef) -> Unit = {},
+): FirValueParameter =
+    coneType.toValueParameter(session, key.origin, containingFunctionSymbol, block)
+
+public fun FirTypeRef.toValueParameter(
+    session: FirSession,
+    origin: FirDeclarationOrigin,
+    containingFunctionSymbol: FirFunctionSymbol<*>,
+    block: FirValueParameterBuilder.(FirTypeRef) -> Unit = {},
+): FirValueParameter = coneType.toValueParameter(session, origin, containingFunctionSymbol, block)
+
+public fun FirFunctionSymbol<*>.contextReceiversToValueParameters(
+    session: FirSession,
+    key: GeneratedDeclarationKey,
+    block: FirValueParameterBuilder.(FirContextReceiver) -> Unit = {},
+): List<FirValueParameter> = contextReceiversToValueParameters(session, key.origin, block)
+
 public fun FirFunctionSymbol<*>.contextReceiversToValueParameters(
     session: FirSession,
     origin: FirDeclarationOrigin,
@@ -46,6 +83,13 @@ public fun FirFunctionSymbol<*>.contextReceiversToValueParameters(
         contextReceiver.toValueParameter(session, origin, this, block)
     }
 }
+
+public fun FirContextReceiver.toValueParameter(
+    session: FirSession,
+    key: GeneratedDeclarationKey,
+    containingFunctionSymbol: FirFunctionSymbol<*>,
+    block: FirValueParameterBuilder.(FirContextReceiver) -> Unit = {},
+): FirValueParameter = toValueParameter(session, key.origin, containingFunctionSymbol, block)
 
 public fun FirContextReceiver.toValueParameter(
     session: FirSession,